Show available, used, and selected leave on cal
An ajax is sent when the calendar page loads to inject available and used leave data for a user into the add event form. Selected event hours are calculated from the event object in javascript.

Creating an event will add the event duration to the pending total for unapproved events and the used total for auto approved events. Approving an event will subtract the event duration from pending and add it to used.

Destroying an event will add back the event duration to pending for unapproved events and add back the duration to used for approved events.