Assign different accrual rate based on years
This commit ands the basic functionality for a manager to create accrual rate rules that determine how much time off an employee accrues based on seniority. Accrual Rules can be set for all types of time off and are blocked for conflicting i.e. one rule 0-10 another 2-5 is not allowed.

This commit does not include updating the nightly accrual chron job to use these rules. The only reason I'm not holding this commit is because I promised a customer these would be delivered today.